+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/service/SubtasksCreationService.java
@@ -154,12 +154,16 @@ public class SubtasksCreationService {
             // try to find provided priority otherwise fall back to priority of parent issue
             if (subTaskRequest.getPriority() != null) {
                 // TODO a priority scheme can be configured per project
-                // TODO add warning for invalid priority
                 Priority priority = priorityManager.getPriorities().stream()
                     .filter(availablePriority -> availablePriority.getName().equals(subTaskRequest.getPriority()))
                     .findFirst()
-                    .orElse(parent.getPriority());
-                newSubtask.setPriority(priority);
+                    .orElse(null);
+                if (priority == null) {
+                    warnings.add("Invalid priority: " + subTaskRequest.getPriority());
+                    newSubtask.setPriority(parent.getPriority());
+                } else {
+                    newSubtask.setPriority(priority);
+                }
             } else {
                 newSubtask.setPriority(parent.getPriority());
             }